requirements.txt
Flask==2.0.2
bcrypt==3.2.0
mysql-connector-python==8.0.27
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
python-dotenv==0.19.2